Separate tests for packaging
